+++ semcore/table/src/index.d.ts
@@ -16,6 +16,8 @@ export interface ITableProps extends IBoxProps {
    * @default primary
    */
   use?: ITableCtx['use'];
+  /** Make cells less */
+  compact?: boolean;
 }
 
 export interface ICellProps extends ITextProps {
@@ -31,6 +33,14 @@ export interface ICellProps extends ITextProps {
    * @default primary
    */
   use?: 'primary' | 'secondary' | false;
+  /** Add vertical border to the right of the cell
+   * @default false
+   */
+  borderRight?: boolean;
+  /** Add vertical border to the left of the cell
+   * @default false
+   */
+  borderLeft?: boolean;
 }
 
 export interface ITableCellHeadProps extends ICellProps {